Recast: upgrade legacy ImGui implementation to 1.90.4 (WIP)
Major upgrade to newer library. This is still work in progress, there are many bugs.

+++ src/naveditor/include/Editor_TileMesh.h
@@ -66,7 +66,7 @@ protected:
 	
 	int m_maxTiles;
 	int m_maxPolysPerTile;
-	float m_tileSize;
+	int m_tileSize;
 	
 	unsigned int m_tileCol;
 	float m_lastBuiltTileBmin[3];
